+++ src/shadow.h
@@ -1,5 +1,5 @@
 #pragma once
 
 struct Shadow {
-    static void Draw(int x, int y, float radius, float yOffset);
+    static void Draw(int x, int y, int radius, int yOffset);
 };
